+++ x/auth/tx/service.go
@@ -115,6 +115,8 @@ func (s txServer) GetTx(ctx context.Context, req *txtypes.GetTxRequest) (*txtype
 		return nil, status.Error(codes.InvalidArgument, "tx hash cannot be empty")
 	}
 
+	// TODO We should also check the proof flag in gRPC header.
+	// https://github.com/cosmos/cosmos-sdk/issues/7036.
 	result, err := QueryTx(s.clientCtx, req.Hash)
 	if err != nil {
 		if strings.Contains(err.Error(), "not found") {